Update README.md (#45)
* Update README.md

The wording before the list of uReplicator goals is clunky.  There is no transition at all between the discussion of MirrorMaker and the goals.  (I am assuming these are the goals, based on identical verbiage at https://github.com/uber/uReplicator/wiki/uReplicator-Design).

This PR simply adds a header to make the transition clear, but the wording is still a bit awkward and could use some proper editorial review.

* Update README.md